requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data & Utilities
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
Pillow>=9.0.0
scipy>=1.7.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Logging & Experiment Tracking
wandb>=0.12.0
tqdm>=4.62.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Optional: For GAN-based perturbations (StarGANv2)
# git+https://github.com/clovaai/stargan2.git
# gdown>=4.0.0

# Optional: For advanced augmentations
# albumentations>=1.0.0
